backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
crewai-tools[mcp]
click
langchain>=0.1.0
langchain-community
opentelemetry-instrumentation-starlette
openinference-instrumentation-crewai